[IMP] membership_extension & _variable_period: Make sure necessary fields are always truthy
This is an additional sanity check to make sure that fields that are
expected to be available are actually available.

The situation here is a little tricky. Especially as pertains date_from
and date_to; these fields are `required=True` nowhere. Not on the
product, not on the membership invoice, and not on the membership line.
However, they _are_ set as required in the product and membership line
views. If these values are ever empty, it's a sure sign that something
went wrong somewhere, because the user should not be able to reach such
a state via the UI.

One sure way to reach that state was via the demo data, however, which
had no start and end dates. This is now fixed.

The situation for the variable period fields is similar. These fields
should never be empty when the membership type is variable, and it's not
typically possible to empty these values via the UI. For convenience's
sake, I simply set them to required an sich, with a default value.

Signed-off-by: Carmen Bianca BAKKER <carmen@carmenbianca.eu>

[BOT] post-merge updates